# encoding/base64: improve performance up to 20% total